requirements.txt
absl-py==1.4.0
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.8.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.2
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
boltons==23.0.0
Bottleneck==1.3.5
Brotli==1.0.9
brotlipy==0.7.0
cachetools==5.3.1
cchardet==2.1.7
certifi==2023.5.7
certipy==0.1.3
cffi==1.15.1
charset-normalizer==2.0.4
chart-studio==1.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
dash==2.6.2
dash-auth==1.4.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.3
Flask==2.1.3
Flask-Compress==1.13
Flask-SeaSurf==1.1.1
flatbuffers==23.5.26
fonttools==4.25.0
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.3
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.54.2
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
intel-cmplr-lib-rt==2023.1.0
intel-cmplr-lic-rt==2023.1.0
intel-opencl-rt==2023.0.0
ipykernel==6.23.1
ipython==8.13.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
Js2Py==0.74
json5==0.9.6
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.2.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyter-telemetry==0.1.0
jupyterhub==4.0.0
jupyterlab==4.0.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
jupyterlab-widgets==3.0.7
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.40.0
lxml==4.9.2
Mako==1.2.3
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.5
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msgpack==1.0.5
multidict==6.0.4
munkres==1.1.4
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.0
numexpr==2.8.4
numpy==1.22.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pip==23.1.2
platformdirs==2.5.2
plotly==5.10.0
pluggy==1.0.0
ply==3.11
pooch==1.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
prosody==0.0.5
protobuf==3.19.6
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.3.0
pycparser==2.21
Pygments==2.15.1
pyjsparser==2.7.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyPrind==2.11.3
pyproject-toml==0.0.10
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.19.3
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
requests==2.29.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
setuptools==67.8.0
sip==6.6.2
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
SQLAlchemy==1.4.39
stack-data==0.6.2
swig==4.1.1
tbb==2021.9.0
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==2.2.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.7.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
ua-parser==0.16.1
uri-template==1.2.0
urllib3==1.26.16
Wave==0.0.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.5
wheel==0.38.4
widgetsnbextension==4.0.7
wrapt==1.15.0
xgboost==1.7.5
yarl==1.9.2
zipp==3.11.0
zstandard==0.19.0
